added auth-middlewares

+++ controllers/todo.js
@@ -1,10 +1,11 @@
 const Todo = require("../models/Todo");
 const mongoose = require('mongoose');
 
-module.exports.getAllTodos = async (req, res) => {
+module.exports.getUserTodos = async (req, res) => {
+  const userId = req.user._id;
 
   try {
-    const todos = await Todo.find();
+    const todos = await Todo.find({ userId });
     res.send({
       todos,
       success: true
@@ -16,8 +17,10 @@ module.exports.getAllTodos = async (req, res) => {
 }
 
 module.exports.createTodo = async (req, res) => {
+  const userId = req.user._id;
   const { title } = req.body;
-  const newTodo = new Todo({title, isDone: false});
+
+  const newTodo = new Todo({ title, isDone: false, userId });
   try {
     await newTodo.save();
     res.send({
@@ -25,30 +28,30 @@ module.exports.createTodo = async (req, res) => {
     });
   }
   catch (err) {
-    res.error(err.toLocaleString());
+    res.status(400).send({ success: false, message: err.toLocaleString() });
   }
 }
 
 module.exports.updateTodo = async (req, res) => {
+  const userId = req.user._id;
   const { todoId } = req.params;
   const { isDone } = req.body;
-  console.log(await Todo.findOne({ _id: mongoose.Types.ObjectId(todoId) })
-  ,isDone)
    try {
-     await Todo.updateOne({ _id: mongoose.Types.ObjectId(todoId) }, { $set: { isDone: isDone } })
-     res.send({
-       success: true
+     await Todo.updateOne({ _id: mongoose.Types.ObjectId(todoId), userId }, { $set: { isDone } })
+       res.send({
+         success: true
      });
    }
    catch (err){
-     res.error(err.toLocaleString());
+     res.code(400).send(err.toLocaleString());
    }
 }
 
 module.exports.deleteTodo = async (req, res) => {
+  const userId = req.user._id;
   const { todoId } = req.params;
   try {
-    await Todo.deleteOne({ _id: mongoose.Types.ObjectId(todoId)});
+    await Todo.deleteOne({ _id: mongoose.Types.ObjectId(todoId), userId });
     res.send({
       success: true
     });
